+++ frontend/package.json
@@ -3,16 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "raphael": "^2.3.0",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
+    "d3": "^7.0.0",
+    "d3.js": "0.0.2-security",
+    "gulp": "^4.0.2",
     "jquery": "^3.6.0",
+    "raphael": "^2.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-svg-pan-zoom": "^3.10.0",
     "react-zoom-pan-pinch": "^1.6.1",
+    "svg-pan-zoom": "^3.6.1",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
